refactor quantization modules
