+++ torchtitan/components/quantization/__init__.py
@@ -1,0 +1,18 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+# [Note] Getting the 'torchao' package:
+# This script requires the 'torchao' package to function correctly.
+# Please ensure you have this package installed from the appropriate repository.
+# You can obtain it from https://github.com/pytorch/ao by following the
+# installation instructions.
+
+# Note: Performance
+# The quantization modules are intended to be ran under `torch.compile`` for competitive performance
+
+# Import to register quantization modules as ModelConverter
+import torchtitan.components.quantization.float8  # noqa: F401
+import torchtitan.components.quantization.mx  # noqa: F401
