+++ torchtitan/components/quantization/float8.py
@@ -4,14 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-# [Note] Getting the 'torchao' package:
-# This script requires the 'torchao' package to function correctly.
-# Please ensure you have this package installed from the appropriate repository.
-# You can obtain it from https://github.com/pytorch/ao by following the
-# installation instructions.
-
-# Note: Performance
-# Float8 experimental is intended to be ran under `torch.compile`` for competitive performance
+from functools import partial
 
 import torch
 import torch.nn as nn
@@ -24,6 +17,8 @@ from torchtitan.protocols.model_converter import (
 )
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability
+
+from .utils import module_filter_fn
 
 
 class Float8Converter(ModelConverter):
@@ -93,12 +88,6 @@ class Float8Converter(ModelConverter):
             logger.info("Float8 tensorwise scaled training active")
 
     def convert(self, model: nn.Module):
-        return self.convert_to_float8_training(model)
-
-    def post_optimizer_hook(self, model: nn.Module | list[nn.Module]):
-        return self.precompute_float8_dynamic_scale_for_fsdp(model)
-
-    def convert_to_float8_training(self, model: nn.Module):
         """
         This function converts the linear layers of `model` to `Float8Linear`.
         Note that today, only dynamic tensor scaling (the default) is supported.
@@ -113,30 +102,14 @@ class Float8Converter(ModelConverter):
         convert_to_float8_training(
             model,
             config=self.config,
-            module_filter_fn=self._module_filter_fn,
+            module_filter_fn=partial(module_filter_fn, filter_fqns=self.filter_fqns),
         )
         logger.info(
             "Swapped to Float8Linear layers with enable_fsdp_float8_all_gather="
             f"{self.config.enable_fsdp_float8_all_gather}"
         )
 
-    def _module_filter_fn(self, mod: nn.Module, fqn: str) -> bool:
-        if not isinstance(mod, nn.Linear):
-            return False
-
-        # All dims must be divisible by 16 due to float8 tensorcore hardware requirements.
-        dims_multiples_of_16 = (
-            mod.weight.shape[0] % 16 == 0 and mod.weight.shape[1] % 16 == 0
-        )
-
-        # If the fqn matches any filtered fqn, then we should not convert this module.
-        is_filtered_fqn = any(filtered_fqn in fqn for filtered_fqn in self.filter_fqns)
-
-        return dims_multiples_of_16 and not is_filtered_fqn
-
-    def precompute_float8_dynamic_scale_for_fsdp(
-        self, model: nn.Module | list[nn.Module]
-    ):
+    def post_optimizer_hook(self, model: nn.Module | list[nn.Module]):
         if not self.enabled:
             return
